refactor: RestController 네이밍 변경
